requirements.txt
streamlit
pandas
pytest
streamlit-tags
fpdf2
Pillow
python-barcode
python-escpos